app_latest/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
pymysql==1.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==1.10.7
requests==2.29.0
python-dotenv==1.0.0
cryptography==40.0.2
pydantic-settings 